src/requirements.txt
pure-python-adb
adb-shell
mysql-connector-python
requests
